# Conflicts:
#	Robot'em Up/Assets/FastIK/Scenes/FastIKSample.unity.meta
#	Robot'em Up/Assets/Mirza Beig/Particle Systems/Ultimate VFX/Demos/Starfall/Starfall.unitypackage.meta
#	Robot'em Up/Assets/ProCore/Polybrush/Shader/z_TextureBlendBmp.shader
#	Robot'em Up/Assets/Scenes/WhiteRooms/WhiteRoom_Quentin_AR.unity.meta
#	Robot'em Up/Assets/Scripts/Player/PawnController.cs

+++ Robot'em Up/Assets/Scripts/IHitable.cs
@@ -5,5 +5,5 @@ public interface IHitable
 	bool lockable_access { get; set; }
 	float lockHitboxSize_access { get; set; }
 
-	void OnHit ( BallBehaviour _ball, Vector3 _impactVector, PawnController _thrower, int _damages, DamageSource _source, Vector3 _bumpModificators = default(Vector3) );
+	void OnHit ( BallBehaviour _ball, Vector3 _impactVector, PawnController _thrower, float _damages, DamageSource _source, Vector3 _bumpModificators = default(Vector3) );
 }
